[util2.message] rename to embed
